refactor: remove agents page
