+++ packages/react-ui/src/features/agents/builder/agent-left-section/agent-prompt-editor.tsx
