+++ packages/react-ui/src/features/agents/builder/agent-structured-output/add-field-popover.tsx
