+++ packages/react-ui/src/features/agents/builder/agent-structured-output/field-type-icon.tsx
