+++ packages/react-ui/src/features/agents/builder/agent-structured-output/index.tsx
